requirements.txt
absl-py==0.1.13
astor==0.6.2
bleach==1.5.0
boto==2.48.0
bz2file==0.98
certifi==2017.7.27.1
chardet==3.0.4
Click==7.0
colorama==0.3.9
cycler==0.10.0
decorator==4.1.2
entrypoints==0.2.3
Flask==1.0.2
gast==0.2.0
gensim==2.3.0
grpcio==1.11.0
h5py==2.8.0
html5lib==0.9999999
idna==2.6
ipykernel==4.6.1
ipython==6.1.0
ipython-genutils==0.2.0
ipywidgets==7.0.0
itsdangerous==0.24
jedi==0.10.2
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.1.0
jupyter-console==5.2.0
jupyter-core==4.3.0
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.1.0rc1
mistune==0.7.4
mmdnn==0.1.3
nbconvert==5.3.0
nbformat==4.4.0
nltk==3.2.5
notebook==5.0.0
numpy==1.15.2+mkl
pandas==0.20.3
pandocfilters==1.4.2
pickleshare==0.7.4
Pillow==5.0.0
prompt-toolkit==1.0.15
protobuf==3.5.2.post1
pyemd==0.4.4
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.2
PyYAML==3.12
pyzmq==16.0.2
qtconsole==4.3.1
requests==2.18.4
scikit-learn==0.19.1
scipy==0.19.1
seaborn==0.8
simplegeneric==0.8.1
six==1.11.0
sklearn==0.0
smart-open==1.5.3
tensorboard==1.7.0
tensorflow-gpu==1.7.0
tensorflow-tensorboard==0.1.5
termcolor==1.1.0
testpath==0.3.1
tornado==4.5.2
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
Werkzeug==0.14.1
widgetsnbextension==3.0.2
win-unicode-console==0.5
